+++ leviathan/profile_taxonomy.py
@@ -1,27 +1,18 @@
 #!/usr/bin/env python
-import sys,os, argparse, warnings, subprocess
-from collections import defaultdict
+import sys
+import os
+import glob
+import pandas as pd
 from tqdm import tqdm
 # from memory_profiler import profile
 
 __program__ = os.path.split(sys.argv[0])[-1]
 
 from pyexeggutor import (
-    # open_file_reader,
-    # open_file_writer,
-    # read_pickle, 
-    # write_pickle,
-    # read_json,
-    # write_json,
-    # build_logger,
-    # reset_logger,
-    # get_timestamp,
-    # format_duration,
-    # format_header,
     format_bytes,
     get_file_size,
-    profile_peak_memory,
     RunShellCommand,
+    check_argument_choice,
 )